+++ src/core/weights.js
