+++ src/explorer/weights/EdgeTypeConfig.js
@@ -4,7 +4,7 @@ import React from "react";
 import {WeightSlider, type Props as WeightSliderProps} from "./WeightSlider";
 
 import type {EdgeType} from "../../analysis/types";
-import type {EdgeWeight} from "../../analysis/weights";
+import type {EdgeWeight} from "../../core/weights";
 
 export class EdgeTypeConfig extends React.Component<{
   +weight: EdgeWeight,
